Added validation logic to Mediator activity classes + necessary changes in ActivitiesContoller

+++ Application/Activities/Delete.cs
@@ -1,4 +1,5 @@
-﻿using Domain;
+﻿using Application.Core;
+using Domain;
 using MediatR;
 using Persistence;
 
@@ -6,11 +7,11 @@ namespace Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>?>
         {
             public Guid Id { get; set; }
 
-            public class Handler : IRequestHandler<Command>
+            public class Handler : IRequestHandler<Command, Result<Unit>?>
             {
                 private readonly DataContext _context;
                 public Handler(DataContext context)
@@ -18,13 +19,20 @@ namespace Application.Activities
                     _context = context;
                 }
 
-                public async Task Handle(Command request, CancellationToken cancellationToken)
+                public async Task<Result<Unit>?> Handle(Command request, CancellationToken cancellationToken)
                 {
-                    Activity activity = await _context.Activities.FindAsync(request.Id);
+                    Activity? activity = await _context.Activities.FindAsync(request.Id);
+
+                    if (activity == null)
+                        return null;
 
                     _context.Activities.Remove(activity);
 
-                    await _context.SaveChangesAsync(cancellationToken);
+                    bool result = await _context.SaveChangesAsync(cancellationToken) > 0;
+
+                    if (!result) return Result<Unit>.Failure("Failed to delete the activity");
+
+                    return Result<Unit>.Success(Unit.Value);
                 }
             }
         }
